requirements.txt
absl-py==0.7.1
astor==0.7.1
Django==2.2
django-crispy-forms==1.7.2
gast==0.2.2
grpcio==1.19.0
h5py==2.9.0
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
Markdown==3.1
mock==2.0.0
numpy==1.16.2
opencv-python==4.0.1.24
pbr==5.1.3
Pillow==6.0.0
protobuf==3.7.1
pytz==2018.9
PyYAML==5.1
scipy==1.2.1
six==1.12.0
sqlparse==0.3.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
Werkzeug==0.15.2
